+++ site/helpers/route.php
@@ -10,9 +10,7 @@
 // no direct access
 defined('_JEXEC') or die;
 
-// Component Helper
-jimport('joomla.application.component.helper');
-jimport('joomla.application.categories');
+jimport("crowdfunding.init");
 
 /**
  * Component Route Helper that help to find a menu item.